update `packages/emotion` tests
