+++ packages/emotion/src/createGlobalStyle.test.tsx
@@ -1,15 +1,10 @@
 import * as React from 'react'
 import '@testing-library/jest-dom/extend-expect'
-import { render, cleanup } from '@testing-library/react'
-import { createGlobalStyle, ThemeProvider } from '.'
+import { cleanup } from '@testing-library/react'
+import { createGlobalStyle } from '.'
+import { renderWithTheme } from './theme.test'
 
 afterEach(cleanup)
-
-const SpaceTheme = ({ children }: { children: React.ReactNode }) => {
-  return (
-    <ThemeProvider theme={{ space: { 1: 4, 2: 8 } }}>{children}</ThemeProvider>
-  )
-}
 
 describe('#createGlobalStyle', () => {
   it('injects global styles', () => {
@@ -18,12 +13,10 @@ describe('#createGlobalStyle', () => {
         margin: 2;
       } 
     `
-    const { container } = render(
+    const { container } = renderWithTheme(
       <>
-        <SpaceTheme>
-          <GlobalStyle />
-          <div className="margin" />
-        </SpaceTheme>
+				<GlobalStyle />
+				<div className="margin" />
       </>,
     )
     expect(container.firstChild).toHaveStyle(`
