+++ packages/emotion/src/css.test.tsx
@@ -1,20 +1,13 @@
 import * as React from 'react'
 import '@testing-library/jest-dom/extend-expect'
-import { render, cleanup } from '@testing-library/react'
+import { cleanup } from '@testing-library/react'
 import styled from '@emotion/styled'
-import { css, ThemeProvider } from '.'
+import { css } from '.'
+import { renderWithTheme } from './theme.test'
 
 afterEach(cleanup)
 
 describe('#css', () => {
-  const SpaceTheme = ({ children }: { children: React.ReactNode }) => {
-    return (
-      <ThemeProvider theme={{ space: { 1: 4, 2: 8 } }}>
-        {children}
-      </ThemeProvider>
-    )
-  }
-
   it('transforms rules', () => {
     const Dummy = styled.div`
       ${css`
@@ -23,11 +16,7 @@ describe('#css', () => {
         margin-top: 2px;
       `}
     `
-    const { container } = render(
-      <SpaceTheme>
-        <Dummy />
-      </SpaceTheme>,
-    )
+    const { container } = renderWithTheme(<Dummy />)
     expect(container.firstChild).toHaveStyle(`
       margin: 2px 8px 8px 8px;
       padding: 4px;
@@ -40,11 +29,7 @@ describe('#css', () => {
         margin: 1 2;
       `}
     `
-    const { container } = render(
-      <SpaceTheme>
-        <Dummy />
-      </SpaceTheme>,
-    )
+    const { container } = renderWithTheme(<Dummy />)
     expect(container.firstChild).toHaveStyle('margin: 4px 8px;')
   })
 
@@ -55,11 +40,7 @@ describe('#css', () => {
         margin: 1 ${two};
       `}
     `
-    const { container } = render(
-      <SpaceTheme>
-        <Dummy />
-      </SpaceTheme>,
-    )
+    const { container } = renderWithTheme(<Dummy />)
     expect(container.firstChild).toHaveStyle('margin: 4px 8px;')
   })
 
@@ -69,25 +50,17 @@ describe('#css', () => {
         margin: '1 2',
       })}
     `
-    const { container } = render(
-      <SpaceTheme>
-        <Dummy />
-      </SpaceTheme>,
-    )
+    const { container } = renderWithTheme(<Dummy />)
     expect(container.firstChild).toHaveStyle('margin: 4px 8px;')
   })
 
   it('accepts function', () => {
     const Dummy = styled.div`
-      ${css((p) => ({
+      ${css((_p) => ({
         margin: '1 2',
       }))}
     `
-    const { container } = render(
-      <SpaceTheme>
-        <Dummy />
-      </SpaceTheme>,
-    )
+    const { container } = renderWithTheme(<Dummy />)
     expect(container.firstChild).toHaveStyle('margin: 4px 8px;')
   })
 
@@ -99,11 +72,7 @@ describe('#css', () => {
         styles: 'margin: 1 2;label:x;',
       })}
     `
-    const { container } = render(
-      <SpaceTheme>
-        <Dummy />
-      </SpaceTheme>,
-    )
+    const { container } = renderWithTheme(<Dummy />)
     expect(container.firstChild).toHaveStyle('margin: 4px 8px;')
   })
 })
