+++ packages/emotion/src/jsx.test.tsx
@@ -1,35 +1,27 @@
 /** @jsx jsx */
 import '@testing-library/jest-dom/extend-expect'
 import { render, cleanup } from '@testing-library/react'
-import { ThemeProvider } from '@emotion/react'
 import { jsx, css } from '.'
+import { renderWithTheme } from './theme.test'
 
 afterEach(cleanup)
-
-const SpaceTheme = ({ children }: { children: React.ReactNode }) => {
-  return (
-    <ThemeProvider theme={{ space: { 1: 4, 2: 8 } }}>{children}</ThemeProvider>
-  )
-}
 
 describe('#jsx', () => {
   it('does nothing without css prop', () => {
     const { container } = render(<div />)
-    expect(container.firstChild!.nodeName).toBe('DIV')
+    expect(container.firstChild?.nodeName).toBe('DIV')
   })
 
   it('handles css string', () => {
-    const { container } = render(
-      <SpaceTheme>
-        <div
-          // @ts-expect-error
-          css={css`
-            margin: 2;
-            padding: 1;
-            margin-top: 2px;
-          `}
-        />
-      </SpaceTheme>,
+    const { container } = renderWithTheme(
+			<div
+				// @ts-expect-error
+				css={css`
+					margin: 2;
+					padding: 1;
+					margin-top: 2px;
+				`}
+			/>
     )
     expect(container.firstChild).toHaveStyle(`
       margin: 2px 8px 8px 8px;
@@ -38,20 +30,18 @@ describe('#jsx', () => {
   })
 
   it('handles array of css string', () => {
-    const { container } = render(
-      <SpaceTheme>
-        <div
-          // @ts-expect-error
-          css={[
-            css`
-              margin: 2;
-            `,
-            css`
-              padding: 1;
-            `,
-          ]}
-        />
-      </SpaceTheme>,
+    const { container } = renderWithTheme(
+			<div
+				// @ts-expect-error
+				css={[
+					css`
+						margin: 2;
+					`,
+					css`
+						padding: 1;
+					`,
+				]}
+			/>
     )
     expect(container.firstChild).toHaveStyle(`
       margin: 8px;
@@ -60,13 +50,11 @@ describe('#jsx', () => {
   })
 
   it('handles css objects', () => {
-    const { container } = render(
-      <SpaceTheme>
-        <div
-          // @ts-expect-error
-          css={{ margin: '2' }}
-        />
-      </SpaceTheme>,
+    const { container } = renderWithTheme(
+			<div
+				// @ts-expect-error
+				css={{ margin: '2' }}
+			/>
     )
     expect(container.firstChild).toHaveStyle(`
       margin: 8px;
@@ -74,13 +62,11 @@ describe('#jsx', () => {
   })
 
   it('handles array of css objects', () => {
-    const { container } = render(
-      <SpaceTheme>
-        <div
-          // @ts-expect-error
-          css={[{ margin: '2' }, { padding: '1' }]}
-        />
-      </SpaceTheme>,
+    const { container } = renderWithTheme(
+			<div
+				// @ts-expect-error
+				css={[{ margin: '2' }, { padding: '1' }]}
+			/>
     )
     expect(container.firstChild).toHaveStyle(`
       margin: 8px;
@@ -89,28 +75,24 @@ describe('#jsx', () => {
   })
 
   it('does not render children', () => {
-    const { container } = render(
-      <SpaceTheme>
-        <div
-          // @ts-expect-error
-          css={[{ margin: '2' }, { padding: '1' }]}
-        />
-      </SpaceTheme>,
+    const { container } = renderWithTheme(
+			<div
+				// @ts-expect-error
+				css={[{ margin: '2' }, { padding: '1' }]}
+			/>
     )
 
     expect(container).toHaveTextContent('')
   })
 
   it('renders a single child', () => {
-    const { container } = render(
-      <SpaceTheme>
-        <div
-          // @ts-expect-error
-          css={[{ margin: '2' }, { padding: '1' }]}
-        >
-          <p id="test-p">A testing paragraph</p>
-        </div>
-      </SpaceTheme>,
+    const { container } = renderWithTheme(
+			<div
+				// @ts-expect-error
+				css={[{ margin: '2' }, { padding: '1' }]}
+			>
+				<p id="test-p">A testing paragraph</p>
+			</div>
     )
 
     expect(container.querySelector('#test-p')).toHaveTextContent(
@@ -119,17 +101,15 @@ describe('#jsx', () => {
   })
 
   it('renders multiple children', () => {
-    const { container } = render(
-      <SpaceTheme>
-        <div
-          // @ts-expect-error
-          css={[{ margin: '2' }, { padding: '1' }]}
-        >
-          <p className="test-p">First testing paragraph</p>
+    const { container } = renderWithTheme(
+			<div
+				// @ts-expect-error
+				css={[{ margin: '2' }, { padding: '1' }]}
+			>
+				<p className="test-p">First testing paragraph</p>
 
-          <p className="test-p">Second testing paragraph</p>
-        </div>
-      </SpaceTheme>,
+				<p className="test-p">Second testing paragraph</p>
+			</div>
     )
 
     expect(container.querySelectorAll('.test-p')).toHaveLength(2)
