+++ packages/emotion/src/theme.test.tsx
@@ -1,14 +1,98 @@
 import * as React from 'react'
 import '@testing-library/jest-dom/extend-expect'
-import { render, cleanup } from '@testing-library/react'
+import { render, cleanup, RenderOptions, RenderResult } from '@testing-library/react'
 import { ThemeProvider } from '@emotion/react'
 import { useTheme, useFontSize } from '.'
+
+// import original module declarations
+import '@xstyled/system'
+import '@emotion/react'
+import {
+	defaultTheme as xstyledDefaultTheme, 
+	DefaultTheme as XStyledDefaultTheme, 
+	th
+} from '@xstyled/system'
+
+import { 
+	ITheme, 
+} from '@xstyled/emotion'
+
+interface AppTheme extends ITheme, XStyledDefaultTheme {
+	fontSizes: XStyledDefaultTheme['fontSizes'] & {
+		md: string
+	}
+	colors: XStyledDefaultTheme['colors'] & {
+		primary: string
+		danger: any
+		text: any
+		red: string
+		modes: { 
+			[key: string]: Partial<AppTheme['colors']>
+		}
+	}
+	space: XStyledDefaultTheme['space'] & {
+		md: number
+	}
+}
+
+// and extend them!
+declare module '@xstyled/system' {
+  // eslint-disable-next-line @typescript-eslint/no-empty-interface
+  export interface Theme extends AppTheme {}
+}
+
+declare module '@emotion/react' {
+  // eslint-disable-next-line @typescript-eslint/no-empty-interface
+  export interface Theme extends AppTheme {
+    /* Customize your theme */
+  }
+}
+
+const defaultTheme: XStyledDefaultTheme = xstyledDefaultTheme
+
+export const theme: AppTheme = {
+	...defaultTheme,
+
+	defaultColorModeName: 'dark',
+	colors: {
+		...defaultTheme.colors,
+		primary: 'pink',
+		black: '#000',
+		white: '#fff',
+		red: '#ff0000',
+		danger: th.color('red'),
+		text: th.color('black'),
+		modes: {
+			dark: {
+				red: '#ff4400',
+				text: th.color('white'),
+			},
+		},
+	},
+	space: {
+		...defaultTheme.space,
+		md: 10,
+		1: '4px', 
+		2: '8px'
+	},
+	fontSizes: { 
+		...defaultTheme.fontSizes,
+		md: '20px',
+	},
+} as const
+
+export const renderWithTheme = (
+	ui: React.ReactElement, 
+	options?: Omit<RenderOptions, 'queries'>
+): RenderResult => render(
+	<ThemeProvider theme={theme}>{ui}</ThemeProvider>,
+	options
+)
 
 afterEach(cleanup)
 
 describe('#useTheme', () => {
   it('returns theme', () => {
-    const theme = { foo: 'bar' }
     const spy = jest.fn()
     function ThemeLogger() {
       const theme = useTheme()
@@ -17,10 +101,8 @@ describe('#useTheme', () => {
       }, [theme])
       return null
     }
-    render(
-      <ThemeProvider theme={theme}>
-        <ThemeLogger />
-      </ThemeProvider>,
+    renderWithTheme(
+			<ThemeLogger />
     )
     expect(spy).toHaveBeenCalledWith(theme)
   })
@@ -28,7 +110,6 @@ describe('#useTheme', () => {
 
 describe('#useFontSize', () => {
   it('gets value from theme', () => {
-    const theme = { fontSizes: { md: '20px' } }
     const spy = jest.fn()
     function Logger() {
       const fontSize = useFontSize('md')
@@ -37,10 +118,8 @@ describe('#useFontSize', () => {
       }, [fontSize])
       return null
     }
-    render(
-      <ThemeProvider theme={theme}>
-        <Logger />
-      </ThemeProvider>,
+    renderWithTheme(
+			<Logger />
     )
     expect(spy).toHaveBeenCalledWith('20px')
   })
